apps/demo-app/requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
redis==5.0.1
prometheus-client==0.19.0
